+++ utils/fileState.js
@@ -6,12 +6,14 @@ const state = {
   CONVERTING: 30, // converting to flac
   UPLOADING: 31, // uploading to server
   WAITING: 32, // in queue waiting to upload
+  STOPPED: 33, // stop gueued file to upload
   PROCESSING: 40, // uploaded, but in process of a.k.a. verifing / ingesting
   COMPLETED: 50 // uploaded and ingested
 }
 
 const preparedGroup = [state.PREPARING, state.ERROR_LOCAL]
 const queuedGroup = [state.WAITING, state.UPLOADING, state.CONVERTING]
+const stoppedGroup = [state.STOPPED]
 const completedGroup = [state.COMPLETED, state.PROCESSING, state.ERROR_SERVER]
 
 const cannotRedoGroup = ['duplicate', 'corrupt', 'duration']
@@ -51,6 +53,7 @@ const getStatePriority = function (id) {
     case state.WAITING: return 3
     case state.PROCESSING: return 4
     case state.COMPLETED: return 7
+    default: return 8
   }
 }
 
@@ -70,6 +73,7 @@ const getName = function (id, message) {
     case state.CONVERTING: return 'compressing'
     case state.WAITING: return 'waiting'
     case state.UPLOADING: return 'uploading'
+    case state.STOPPED: return 'stopped'
     case state.PROCESSING: return 'verified'
     case state.ERROR_LOCAL:
     case state.ERROR_SERVER:
@@ -90,6 +94,8 @@ const getIconName = function (id) {
       return 'ic-state-uploading.svg'
     case state.PROCESSING:
       return 'ic-state-ingesting.svg'
+    case state.STOPPED:
+      return 'ic-state-waiting.svg'
     case state.ERROR_LOCAL:
     case state.ERROR_SERVER:
     case state.DUPLICATED:
@@ -104,6 +110,10 @@ const isInPreparedGroup = function (state) {
 
 const isInQueuedGroup = function (state) {
   return queuedGroup.includes(state)
+}
+
+const isInStoppedGroup = function (state) {
+  return stoppedGroup.includes(state)
 }
 
 const isInCompletedGroup = function (state) {
@@ -161,6 +171,7 @@ export default {
   getStateByStateId,
   preparedGroup,
   queuedGroup,
+  stoppedGroup,
   completedGroup,
   getStatePriority,
   getSummaryStatePriority,
@@ -177,6 +188,7 @@ export default {
   isProcessing,
   isDuplicated,
   isCompleted,
+  isInStoppedGroup,
   cannotRedoGroup,
   canRedo,
   canRemove,